requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-texttospeech==2.15.4
google-cloud-speech==2.26.0
